// packages/three-vrm/src/types/GLTFSchema.ts
export namespace GLTFSchema {
// ...
  export interface Material {
    /**
     * The alpha cutoff value of the material.
     */
    alphaCutoff?: number;
    /**
     * The alpha rendering mode of the material.
     */
    alphaMode?: string;
    /**
     * Specifies whether the material is double sided.
     */
    doubleSided?: boolean;
    /**
     * The emissive color of the material.
     */
    emissiveFactor?: number[];
    /**
     * The emissive map texture.
     */
    emissiveTexture?: TextureInfo;
    extensions?: { [key: string]: { [key: string]: any } };
    extras?: any;
    /**
     * The user-defined name of this object.
     */
    name?: string;
    /**
     * The normal map texture.
     */
    normalTexture?: MaterialNormalTextureInfoObject;
    /**
     * The occlusion map texture.
     */
    occlusionTexture?: MaterialOcclusionTextureInfoObject;
    /**
     * A set of parameter values that are used to define the metallic-roughness material model
     * from Physically-Based Rendering (PBR) methodology. When not specified, all the default
     * values of `pbrMetallicRoughness` apply.
     */
    pbrMetallicRoughness?: MaterialPBRMetallicRoughness;
  }

  /**
   * The emissive map texture.
   *
   * The base color texture.
   *
   * The metallic-roughness texture.
   *
   * Reference to a texture.
   */
  export interface TextureInfo {
    extensions?: { [key: string]: { [key: string]: any } };
    extras?: any;
    /**
     * The index of the texture.
     */
    index: number;
    /**
     * The set index of texture's TEXCOORD attribute used for texture coordinate mapping.
     */
    texCoord?: number;
  }

  /**
   * The normal map texture.
   *
   * Reference to a texture.
   */
  export interface MaterialNormalTextureInfoObject {
    extensions?: { [key: string]: { [key: string]: any } };
    extras?: any;
    /**
     * The index of the texture.
     */
    index: number;
    /**
     * The scalar multiplier applied to each normal vector of the normal texture.
     */
    scale?: number;
    /**
     * The set index of texture's TEXCOORD attribute used for texture coordinate mapping.
     */
    texCoord?: number;
  }

  /**
   * The occlusion map texture.
   *
   * Reference to a texture.
   */
  export interface MaterialOcclusionTextureInfoObject {
    extensions?: { [key: string]: { [key: string]: any } };
    extras?: any;
    /**
     * The index of the texture.
     */
    index: number;
    /**
     * A scalar multiplier controlling the amount of occlusion applied.
     */
    strength?: number;
    /**
     * The set index of texture's TEXCOORD attribute used for texture coordinate mapping.
     */
    texCoord?: number;
  }

  /**
   * A set of parameter values that are used to define the metallic-roughness material model
   * from Physically-Based Rendering (PBR) methodology. When not specified, all the default
   * values of `pbrMetallicRoughness` apply.
   *
   * A set of parameter values that are used to define the metallic-roughness material model
   * from Physically-Based Rendering (PBR) methodology.
   */
  export interface MaterialPBRMetallicRoughness {
    /**
     * The material's base color factor.
     */
    baseColorFactor?: number[];
    /**
     * The base color texture.
     */
    baseColorTexture?: TextureInfo;
    extensions?: { [key: string]: { [key: string]: any } };
    extras?: any;
    /**
     * The metalness of the material.
     */
    metallicFactor?: number;
    /**
     * The metallic-roughness texture.
     */
    metallicRoughnessTexture?: TextureInfo;
    /**
     * The roughness of the material.
     */
    roughnessFactor?: number;
  }
// ...
  export interface Node {
    /**
     * The index of the camera referenced by this node.
     */
    camera?: number;
    /**
     * The indices of this node's children.
     */
    children?: number[];
    extensions?: { [key: string]: { [key: string]: any } };
    extras?: any;
    /**
     * A floating-point 4x4 transformation matrix stored in column-major order.
     */
    matrix?: number[];
    /**
     * The index of the mesh in this node.
     */
    mesh?: number;
    /**
     * The user-defined name of this object.
     */
    name?: string;
    /**
     * The node's unit quaternion rotation in the order (x, y, z, w), where w is the scalar.
     */
    rotation?: number[];
    /**
     * The node's non-uniform scale, given as the scaling factors along the x, y, and z axes.
     */
    scale?: number[];
    /**
     * The index of the skin referenced by this node.
     */
    skin?: number;
    /**
     * The node's translation along the x, y, and z axes.
     */
    translation?: number[];
    /**
     * The weights of the instantiated Morph Target. Number of elements must match number of
     * Morph Targets of used mesh.
     */
    weights?: number[];
  }
// ...
}